Added end-to-end tests for ranged for loops
Turns out you have to explicitly append `f` to the end of floats, so doubles are a more reasonable cross-language floating number. Fine.

Fixes Ruby's "continue" to be "next".

+++ test/integration/IfStringToFloatEnd/if string to float end.java
@@ -1,8 +1,8 @@
 //
-float bbb = null;
+double bbb = null;
 
 try {
-    bbb = Float.parseFloat(aaa);
+    bbb = Double.parseDouble(aaa);
 } catch (NumberFormatException e) { }
 
 if (bbb != null) {
